refactor(ffi): add #[link_name] and delete C pass-through wrappers (Phase 1)
Delete 4 C pass-through wrapper functions by adding `#[link_name]`
attributes to their Rust extern declarations, allowing Rust to link
directly to the real C symbols:

- `nvim_xstrdup` → `xstrdup` (memory_ffi.c, 19 Rust files)
- `nvim_curbufIsChanged` → `curbufIsChanged` (change_ffi.c, 3 files)
- `nvim_ml_get_len` → `ml_get_len` (change_ffi.c, 8 files)
- `nvim_utf_char2bytes` → `utf_char2bytes` (change_ffi.c, 3 files)

Handle clashing extern declarations caused by existing direct `fn`
declarations of the same C symbol in sibling modules:
- buffer/src/info.rs: switch nvim_curbufIsChanged → curbufIsChanged() -> bool
- option/src/value.rs: same switch
- edit/src/putchar.rs: switch nvim_utf_char2bytes → direct utf_char2bytes with *mut u8
- memline/src/recovery.rs: fix return type i64 → i32 to match colnr_T

+++ src/nvim-rs/memline/src/recovery.rs
@@ -221,7 +221,8 @@ extern "C" {
     fn nvim_set_fenc_local(fenc: *const c_char);
     fn nvim_ml_append_recovery(lnum: i64, line: *const c_char, is_new: bool) -> c_int;
     fn nvim_ml_get(lnum: i64) -> *mut c_char;
-    fn nvim_ml_get_len(lnum: i64) -> i64;
+    #[link_name = "ml_get_len"]
+    fn nvim_ml_get_len(lnum: i64) -> i32;
     #[link_name = "xstrnsave"]
     fn nvim_xstrnsave(s: *const c_char, len: usize) -> *mut c_char;
     #[link_name = "buf_inc_changedtick"]
